requirements.txt
setuptools
streamlit
numpy
Pillow
tensorflow
matplotlib
seaborn
pandas
scikit-learn
